feat(dragdrop): Multi target the Window.RootElement property
