+++ src/Uno.UI/UI/Xaml/Window.Android.cs
@@ -70,10 +70,9 @@ namespace Windows.UI.Xaml
 			_rootBorder.Child = _content = value;
 		}
 
-		private UIElement InternalGetContent()
-		{
-			return _content;
-		}
+		private UIElement InternalGetContent() => _content;
+
+		private UIElement InternalGetRootElement() => _main;
 
 		internal UIElement MainContent => _main;
 
@@ -244,7 +243,7 @@ namespace Windows.UI.Xaml
 			}
 		}
 
-		#region StatusBar properties
+#region StatusBar properties
 		private bool IsStatusBarVisible()
 		{
 				var decorView = (ContextHelper.Current as Activity)?.Window?.DecorView;
@@ -267,9 +266,9 @@ namespace Windows.UI.Xaml
 			return activity.Window.Attributes.Flags.HasFlag(WindowManagerFlags.TranslucentStatus)
 				|| activity.Window.Attributes.Flags.HasFlag(WindowManagerFlags.LayoutNoLimits); ;
 		}
-		#endregion
+#endregion
 
-		#region NavigationBar properties
+#region NavigationBar properties
 		private bool IsNavigationBarVisible()
 		{
 			var decorView = (ContextHelper.Current as Activity)?.Window?.DecorView;
@@ -294,7 +293,7 @@ namespace Windows.UI.Xaml
 			return flags.HasFlag(WindowManagerFlags.TranslucentNavigation)
 				|| flags.HasFlag(WindowManagerFlags.LayoutNoLimits);
 		}
-		#endregion
+#endregion
 
 		internal IDisposable OpenPopup(Popup popup)
 		{
