+++ src/Uno.UI/UI/Xaml/Window.iOS.cs
@@ -116,6 +116,8 @@ namespace Windows.UI.Xaml
 
 		private UIElement InternalGetContent() => _content;
 
+		private UIElement InternalGetRootElement() => _main;
+
 		private static Window InternalGetCurrentWindow()
 		{
 			if (_current == null)
