+++ src/Uno.UI/UI/Xaml/Window.macOS.cs
@@ -103,6 +103,8 @@ namespace Windows.UI.Xaml
 
 		private UIElement InternalGetContent() => _content;
 
+		private UIElement InternalGetRootElement() => _main;
+
 		private static Window InternalGetCurrentWindow()
 		{
 			if (_current == null)
